% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vague}
\alias{vague}
\title{Vague picture}
\usage{
vague(pic)
}
\arguments{
\item{pic}{is the input matrix needs to vague}
}
\value{
the matrix which has already vagued
}
\description{
Vague picture
}
\examples{
require(recoverpic)
vimg = vague(img)
}
